+++ packages/wordpress/CHANGELOG.md
@@ -1,1 +1,5 @@
 ## Unreleased
+
+### Bug Fixes
+
+- Fixed a bug in the cache mechanism affecting the Render block. [[[🔗 Bug]](https://community.blockera.ai/changelog-9l8hbrv0/post/version-1-0---in-development-oz0Mrh3r3JN0QDO)]